# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Exam Backend API",
    "description": "API documentation for Sindre Gangeskar's Exam Project",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "adminAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Include **Bearer** prefix in the value field with token - e.g **Bearer abc123**"
    },
    "userAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Include **Bearer** prefix in the value field with token - e.g **Bearer abc123**"
    }
  },
  "paths": {
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Any user can view all available brands",
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully retrieves all brands. If no brands are found, an **empty array** will be returned",
            "schema": {
              "$ref": "#/definitions/brandsGetSuccess"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API is trying to retrieve all available brands",
            "schema": {
              "$ref": "#/definitions/brandsGetInternalError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "An **admin** can create a new brand for products to use",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "The name property is **required** in order to create a brand",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Example Brand"
                }
              },
              "required": [
                "name"
              ]
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully creates a new brand",
            "schema": {
              "$ref": "#/definitions/brandsPostSuccess"
            }
          },
          "400": {
            "description": "This response is returned when the **admin** user provides incorrect or invalid values or properties in the body",
            "schema": {
              "$ref": "#/definitions/brandsPostUserError"
            }
          },
          "401": {
            "description": "This response is returned when the user lacks **admin** privileges or isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/adminUnauthorized"
            }
          },
          "409": {
            "description": "This response is returned when the API tries to create a brand, but another brand with the provided brand name already exists",
            "schema": {
              "$ref": "#/definitions/brandsPostConflict"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API is trying to create a new brand",
            "schema": {
              "$ref": "#/definitions/brandsPostInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          }
        ]
      }
    },
    "/brands/{id}": {
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "An **admin** can delete a brand",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The brand id"
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully deletes a brand",
            "schema": {
              "$ref": "#/definitions/brandsDeleteSuccess"
            }
          },
          "401": {
            "description": "This response is returned when the user lacks **admin** privileges or isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/adminUnauthorized"
            }
          },
          "404": {
            "description": "This response is returned when the API fails to find a brand with the provided id ",
            "schema": {
              "$ref": "#/definitions/brandsDeleteNotFound"
            }
          },
          "409": {
            "description": "This response is returned when the API fails to delete the brand because a product or several are using the brand **admin** wants to delete",
            "schema": {
              "$ref": "#/definitions/brandsDeleteConflict"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API attempts to delete a brand",
            "schema": {
              "$ref": "#/definitions/brandsDeleteInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "An **admin** can update a brand\\'s name",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The brand id"
          },
          {
            "name": "body",
            "in": "body",
            "description": "The **name** property is required in order to update a brand's name",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Another brand name"
                }
              },
              "required": [
                "name"
              ]
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully updates a brand",
            "schema": {
              "$ref": "#/definitions/brandsUpdateSuccess"
            }
          },
          "400": {
            "description": "This response is returned when **admin** provides an invalid value",
            "schema": {
              "$ref": "#/definitions/brandsUpdateUserError"
            }
          },
          "401": {
            "description": "This response is returned when the user lacks **admin** privileges or isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/adminUnauthorized"
            }
          },
          "409": {
            "description": "This response is returned when there's a conflict while trying to update the brand's name to something another brand is already using",
            "schema": {
              "$ref": "#/definitions/brandsUpdateConflict"
            }
          },
          "500": {
            "description": "This response is returned when an internal server occurrs while the API is trying to update brand",
            "schema": {
              "$ref": "#/definitions/brandsUpdateInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "adminUnauthorized": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 401
            },
            "result": {
              "type": "array",
              "example": [
                "Login required",
                "Admin privileges required",
                "Invalid token, please log in again",
                "Session has expired, please log in again"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "InvalidJWTError",
                "InvalidCredentialsError",
                "UnauthorizedError",
                "TokenExpiredError"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "brandsGetSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "array",
              "example": [
                "Successfully found brands",
                "No brands found"
              ],
              "items": {
                "type": "string"
              }
            },
            "brands": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Apple"
                  }
                }
              }
            }
          }
        }
      }
    },
    "brandsGetInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "message": {
              "type": "string",
              "example": "An internal server error has occurred while trying to get all brands"
            }
          }
        }
      }
    },
    "brandsDeleteSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully deleted brand"
            }
          }
        }
      }
    },
    "brandsDeleteNotFound": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 404
            },
            "result": {
              "type": "string",
              "example": "Cannot find brand with provided id"
            },
            "error": {
              "type": "string",
              "example": "BrandNotFoundError"
            }
          }
        }
      }
    },
    "brandsDeleteConflict": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 409
            },
            "result": {
              "type": "string",
              "example": "Cannot delete brand - A product is using this brand"
            },
            "error": {
              "type": "string",
              "example": "BrandDeleteError"
            }
          }
        }
      }
    },
    "brandsDeleteInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "message": {
              "type": "string",
              "example": "An internal server error has occurred while trying to delete brand"
            }
          }
        }
      }
    },
    "brandsPostSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully created brand"
            }
          }
        }
      }
    },
    "brandsPostUserError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 400
            },
            "result": {
              "type": "array",
              "example": [
                "Required properties are missing",
                "Please remove invalid properties",
                "name must be a non-empty string"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "MissingPropertyError",
                "InvalidPropertyError",
                "InvalidStringError"
              ],
              "items": {
                "type": "string"
              }
            },
            "missingInBody": {
              "type": "array",
              "example": [
                "name"
              ],
              "items": {
                "type": "string"
              }
            },
            "invalidProperties": {
              "type": "array",
              "example": [
                "ExampleInvalidProperty1",
                "ExampleInvalidProperty2"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "brandsPostInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "message": {
              "type": "string",
              "example": "An internal server error has occurred while trying to create brand"
            }
          }
        }
      }
    },
    "brandsPostConflict": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 409
            },
            "result": {
              "type": "string",
              "example": "Brand with provided name already exists"
            },
            "error": {
              "type": "string",
              "example": "DuplicateRecordError"
            }
          }
        }
      }
    },
    "brandsUpdateSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully updated brand"
            }
          }
        }
      }
    },
    "brandsUpdateUserError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 400
            },
            "result": {
              "type": "array",
              "example": [
                "Please remove invalid properties",
                "At least one property must be present",
                "name must be a non-empty string"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "InvalidPropertyError",
                "BodyEmptyUserError",
                "InvalidStringError"
              ],
              "items": {
                "type": "string"
              }
            },
            "validProperties": {
              "type": "array",
              "example": [
                "name"
              ],
              "items": {
                "type": "string"
              }
            },
            "invalidProperties": {
              "type": "array",
              "example": [
                "InvalidPropertyExample1",
                "InvalidPropertyExample2"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "brandsUpdateConflict": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 409
            },
            "result": {
              "type": "string",
              "example": "Cannot update brand - another brand with the same name already exists"
            },
            "error": {
              "type": "string",
              "example": "DuplicateRecordError"
            }
          }
        }
      }
    },
    "brandsUpdateInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "result": {
              "type": "string",
              "example": "An internal server error has occurred while trying to update brand"
            }
          }
        }
      }
    },
...
  }
}